Added an option to do a full ping request
Technically the 2 bytes it sends by default is enough. But by
enabling this option it'll send data according to the actual protocol.
Only good reason to enable this would be if the server is behind a
mcvhostproxy which forwards these requests. The standard request of 2
bytes would not work. tl;dr, you probably don't need this option. I
support it anyway.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>